Rename src/{ => config}/setupVars.{c,h}
