+++ src/api/dns.c
@@ -13,7 +13,7 @@
 #include "webserver/json_macros.h"
 #include "api.h"
 // {s,g}et_blockingstatus()
-#include "setupVars.h"
+#include "config/setupVars.h"
 // set_blockingmode_timer()
 #include "timers.h"
 #include "shmem.h"
