+++ src/config/CMakeLists.txt
@@ -21,6 +21,8 @@ set(sources
         legacy_reader.h
         password.c
         password.h
+        setupVars.c
+        setupVars.h
         toml_writer.c
         toml_writer.h
         toml_reader.c
