+++ src/dnsmasq_interface.c
@@ -20,7 +20,7 @@
 #include "database/database-thread.h"
 #include "datastructure.h"
 #include "database/gravity-db.h"
-#include "setupVars.h"
+#include "config/setupVars.h"
 #include "daemon.h"
 #include "timers.h"
 #include "gc.h"
